Update encoding_polar_codes.py
Add generator matrix for polar code and then do the Kronecker product of this generator matrix.
Here I am using minimum_kronecker because this is the minimum of Kronecker multiplication. If the length of the codeword is 4 then Kronecker's multiplication should be 2.
Although we are using the codeword_length = 3, then the generator matrix will not do the repetition since minimum_kronecker is already equal with kronecker_multiplier.
For generate the polar codes, I have create the comment and you can check if it is okay or not. Please give me your advice. :D